+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 * Fetching invalid ddl confuguration (sharding key for non-existing space)
   is no longer breaks CRUD requests (#308, PR #309).
+* ddl space record delete no more throws error if crud is used (#310, PR #311).
+* crud sharding metainfo is now updated on ddl record delete (#310, PR #311).
 
 ## [0.12.0] - 28-06-22